Functionality updated and directory structure changed.
