+++ src/main/java/com/example/moviecharactersapi/controllers/CharacterController.java
@@ -1,6 +1,6 @@
 package com.example.moviecharactersapi.controllers;
 
-import com.example.moviecharactersapi.entities.Character;
+import com.example.moviecharactersapi.models.Character;
 import com.example.moviecharactersapi.services.service_interfaces.CharacterService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
