--- src/main/java/com/example/moviecharactersapi/entities/GenderType.java
@@ -1,5 +1,0 @@
-package com.example.moviecharactersapi.entities;
-
-public enum GenderType {
-
-}
