+++ src/main/java/com/example/moviecharactersapi/models/Character.java
@@ -1,0 +1,26 @@
+package com.example.moviecharactersapi.models;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Set;
+
+@Entity
+@Getter
+@Setter
+public class  Character {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    @Column(length=50, nullable = false)
+    private String full_name;
+    @Column(length=50)
+    private String alias;
+    @Enumerated(EnumType.STRING)
+    private GenderType gender;
+    @Column(length = 100)
+    private String picture;
+    @ManyToMany(mappedBy = "characters")
+    private Set<Movie> movies;
+}
