+++ src/main/java/com/example/moviecharactersapi/models/GenderType.java
@@ -1,0 +1,5 @@
+package com.example.moviecharactersapi.models;
+
+public enum GenderType {
+
+}
