+++ src/main/java/com/example/moviecharactersapi/models/Movie.java
@@ -1,0 +1,33 @@
+package com.example.moviecharactersapi.models;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Set;
+
+@Entity
+@Getter
+@Setter
+public class Movie {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    @Column(length = 30, nullable = false)
+    private String title;
+    @Column(length = 50, nullable = false)
+    private String genre;
+    @Column(nullable = false)
+    private int year_of_release;
+    @Column(length = 100, nullable = false)
+    private String director;
+    @Column(length = 100)
+    private String picture;
+    @Column(length = 100)
+    private String trailer;
+    @ManyToOne
+    @JoinColumn(name = "franchise_id")
+    private Franchise franchise;
+    @ManyToMany
+    private Set<Character> characters;
+}
