+++ src/main/java/com/example/moviecharactersapi/repositories/FranchiseRepository.java
@@ -1,6 +1,6 @@
 package com.example.moviecharactersapi.repositories;
 
-import com.example.moviecharactersapi.entities.Franchise;
+import com.example.moviecharactersapi.models.Franchise;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
