+++ src/main/java/com/example/moviecharactersapi/repositories/MovieRepository.java
@@ -1,6 +1,6 @@
 package com.example.moviecharactersapi.repositories;
 
-import com.example.moviecharactersapi.entities.Movie;
+import com.example.moviecharactersapi.models.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
