+++ src/main/java/com/example/moviecharactersapi/services/service_interfaces/CharacterService.java
@@ -1,6 +1,6 @@
 package com.example.moviecharactersapi.services.service_interfaces;
 
-import com.example.moviecharactersapi.entities.Character;
+import com.example.moviecharactersapi.models.Character;
 
 public interface CharacterService extends CrudService<Character, Integer>{
 }
