+++ src/main/java/com/example/moviecharactersapi/services/service_interfaces/FranchiseService.java
@@ -1,6 +1,6 @@
 package com.example.moviecharactersapi.services.service_interfaces;
 
-import com.example.moviecharactersapi.entities.Franchise;
+import com.example.moviecharactersapi.models.Franchise;
 
 public interface FranchiseService extends CrudService<Franchise, Integer> {
 }
