+++ src/main/java/com/example/moviecharactersapi/services/service_interfaces/MovieService.java
@@ -1,6 +1,6 @@
 package com.example.moviecharactersapi.services.service_interfaces;
 
-import com.example.moviecharactersapi.entities.Movie;
+import com.example.moviecharactersapi.models.Movie;
 
 public interface MovieService extends CrudService<Movie, Integer>{
 }
